udev tests: Re-org. code
Break code out that is of use for other udev tests.
